requirements.txt
python-telegram-bot>=20.0
anthropic>=0.18.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0
joblib>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
tensorflow-cpu>=2.16.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.2.0
jinja2>=3.1.0
psutil>=5.9.0
python-i18n>=0.3.9
langdetect>=1.0.9